fix: delete announcement permission

+++ src/controllers/announcement/announcement.controller.ts
@@ -8,6 +8,7 @@ import {
   deleteAnnouncementService,
 } from "../../services";
 import { getAnnouncementByIdService } from "../../services/announcement/getAnnouncementbyId.service";
+import { AppError } from "../../errors";
 
 export const createAnnouncementController = async (
   request: Request,
@@ -81,6 +82,10 @@ export const deleteCarController = async (
   request: Request,
   response: Response
 ): Promise<Response> => {
+  if (response.locals.userId != response.locals.findAnnouncement) {
+    throw new AppError("Insufficient permission", 403);
+  }
+
   const id = request.params.id;
 
   await deleteAnnouncementService(id);
